Update tests, add try/except in main for move config to appdir
